% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coppeR.R
\name{get_coppeR}
\alias{get_coppeR}
\title{Import Pictures of Copper}
\usage{
get_coppeR(x)
}
\description{
This function allows you to get a lovely dose of Copper when R is not working
keywords Copper
}
\examples{
get_coppeR()
}

